+++ packages/loader/src/components/tj-loader/loader.ts
@@ -1,30 +1,23 @@
-
-
-import style from "./loader.scss?inline";
-import {tj_loader_state_internal} from "../../lib/tj-loader-state";
-
+import { ScrollHandler } from '../../lib/scroll-handler';
+import { tj_loader_state_internal } from '../../lib/tj-loader-state';
+import style from './loader.scss?inline';
 
 async function sleep(ms: number) {
-  return new Promise(resolve => setTimeout(resolve, ms));
+  return new Promise((resolve) => setTimeout(resolve, ms));
 }
-
-
-
-
-
 
 const startTime = Date.now();
 
 export class LoaderElement extends HTMLElement {
-
-
   #elementMap = new Map<HTMLElement, { waitStart: number }>();
 
   #startTime = startTime;
 
-  #interval : number | null = null;
+  #interval: number | null = null;
 
   #onAfterLoad = false;
+
+  #scrollHandler: ScrollHandler | null = null;
 
   constructor() {
     super();
@@ -38,16 +31,15 @@ export class LoaderElement extends HTMLElement {
 
     // Find the first Image element in document
 
-
     rootElement.innerHTML = `<div id="wrapper"><slot name="loader"><div id="window"><div id="image"><img src="" loading="eager" fetchpriority="high"></div><div id="loadbar"></div></div></slot></div><slot id="main"></slot>`;
     shadowRoot.appendChild(rootElement);
   }
 
   connectedCallback() {
-    window.tj_loader_state = "loading";
+    window.tj_loader_state = 'loading';
 
-    this.addEventListener('init:child-waitreq', (e) => this.#handleChildWaitReq(e as CustomEvent));
-    this.addEventListener('init:child-ready', (e) => this.#handleChildReady(e as CustomEvent));
+    window.addEventListener('init:child-waitreq', (e) => this.#handleChildWaitReq(e as CustomEvent));
+    window.addEventListener('init:child-ready', (e) => this.#handleChildReady(e as CustomEvent));
 
     this.#interval = window.setInterval(this.#checkReadyState, 2000);
 
@@ -60,18 +52,34 @@ export class LoaderElement extends HTMLElement {
     });
 
     window.setTimeout(() => {
-      const firstImg = document.querySelector("img.loader") ?? document.querySelector("img");
+      const firstImg = document.querySelector('img.loader') ?? document.querySelector('img');
       const imageSrc = firstImg?.getAttribute('src') || this.getAttribute('data-src') || '';
-      const img = this.shadowRoot!.querySelector("img");
+      const img = this.shadowRoot!.querySelector('img');
       if (img) {
         img.onload = () => {
           img.classList.add('loaded');
-        }
-        img.setAttribute("src", imageSrc);
+        };
+        img.setAttribute('src', imageSrc);
       }
-    },2);
+    }, 2);
+  }
 
+  #registerScrollHandler() {
+    const selector = this.getAttribute('observe-scroll-element');
 
+    let scrollElement: Window | HTMLElement | null = window;
+    if (selector) {
+      scrollElement = document.querySelector(selector) as HTMLElement | null;
+      if (!scrollElement) {
+        console.warn(
+          `Scroll handler observe-scroll-element: '${selector}' did not match any element. Scroll restoration will be disabled.`,
+        );
+        return;
+      }
+    }
+    this.#scrollHandler = new ScrollHandler(scrollElement);
+    this.#scrollHandler?.connectEventListener();
+    this.#scrollHandler?.restoreScrollPosition();
   }
 
   #checkReadyState = async () => {
@@ -80,74 +88,76 @@ export class LoaderElement extends HTMLElement {
     const now = Date.now();
     for (const [element, info] of this.#elementMap.entries()) {
       if (now - info.waitStart > 4000) {
-        console.error(`Element ${element} has been waiting for more than 4 seconds. Removing from loader (Check callbacks!).`, element);
+        console.error(
+          `Element ${element} has been waiting for more than 4 seconds. Removing from loader (Check callbacks!).`,
+          element,
+        );
         this.#elementMap.delete(element);
       }
     }
 
-    if ( ! this.#onAfterLoad) {
+    if (!this.#onAfterLoad) {
       return; // Wait for ready stat
     }
 
     if (this.#elementMap.size === 0) {
-
       window.clearInterval(this.#interval!);
 
       this.classList.add('ready');
       await sleep(1); // Ensure ready (display: block) state is applied before firing event
-      tj_loader_state_internal.state = "ready";
-      this.dispatchEvent(new CustomEvent('loader:ready', {
-        bubbles: true,
-        composed: true,
-      }));
+      tj_loader_state_internal.state = 'ready';
+      this.dispatchEvent(
+        new CustomEvent('loader:ready', {
+          bubbles: true,
+          composed: true,
+        }),
+      );
       console.debug(`Loader ready after ${Date.now() - this.#startTime}ms`);
 
-
-
       await sleep(10); // Ensure ready state is applied before visual state
-      tj_loader_state_internal.state = "pre-visual";
+      tj_loader_state_internal.state = 'pre-visual';
       this.classList.add('pre-visual');
-      this.dispatchEvent(new CustomEvent('loader:pre-visual', {
-        bubbles: true,
-        composed: true,
-      }));
+      this.dispatchEvent(
+        new CustomEvent('loader:pre-visual', {
+          bubbles: true,
+          composed: true,
+        }),
+      );
 
       console.debug(`Loader pre-visual after ${Date.now() - this.#startTime}ms`);
 
       await sleep(150); // Ensure ready state is applied before visual state
-      tj_loader_state_internal.state = "visual";
+      tj_loader_state_internal.state = 'visual';
       this.classList.add('visual'); // Ensures no animation is waiting
       await sleep(1); // Ensure ready state is applied before visual state
-      this.dispatchEvent(new CustomEvent('loader:visual', {
-        bubbles: true,
-        composed: true,
-      }));
+      this.dispatchEvent(
+        new CustomEvent('loader:visual', {
+          bubbles: true,
+          composed: true,
+        }),
+      );
 
+      this.#registerScrollHandler();
       console.debug(`Loader visual after ${Date.now() - this.#startTime}ms`);
-
     }
-  }
-
-  #handleChildWaitReq = (event: CustomEvent) => {
-    const {element, state} = event.detail;
-    this.#elementMap.set(element, {waitStart: Date.now()});
   };
 
-
-
+  #handleChildWaitReq = (event: CustomEvent) => {
+    const { element, state } = event.detail;
+    this.#elementMap.set(element, { waitStart: Date.now() });
+  };
 
   #handleChildReady = (event: CustomEvent) => {
-    const {element, state} = event.detail;
+    const { element, state } = event.detail;
     const info = this.#elementMap.get(element);
-    if ( ! info) {
+    if (!info) {
       console.warn(`Received ready event for element that did not send waitreq:`, element);
       return;
     }
     this.#elementMap.delete(element);
     console.debug(`Element ready:`, element, `Waited for ${Date.now() - info.waitStart}ms`);
-    this.#checkReadyState()
-  }
-
+    this.#checkReadyState();
+  };
 }
 
 // check if the element is already defined - if so trigger error