+++ packages/ast-markdown/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.9 (2026-05-15)
+
+This was a version bump only for ast-markdown to align it with other projects, there were no code changes.
+
+## 1.0.8 (2026-05-15)
+
+This was a version bump only for ast-markdown to align it with other projects, there were no code changes.
+
 ## 1.0.7 (2026-02-08)
 
 This was a version bump only for ast-markdown to align it with other projects, there were no code changes.