+++ packages/markdown-loader/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.12 (2026-05-18)
+
+This was a version bump only for markdown-loader to align it with other projects, there were no code changes.
+
+## 1.0.11 (2026-05-18)
+
+This was a version bump only for markdown-loader to align it with other projects, there were no code changes.
+
+## 1.0.10 (2026-05-18)
+
+This was a version bump only for markdown-loader to align it with other projects, there were no code changes.
+
 ## 1.0.9 (2025-11-11)
 
 This was a version bump only for markdown-loader to align it with other projects, there were no code changes.